requirements.txt
django==5.0.6
djangorestframework==3.14.0
django-cors-headers==3.13.0